+++ server/src/server/editorServices.ts
@@ -2623,6 +2623,7 @@ export class ProjectService {
 
     /** @internal */
     applyChangesToFile(scriptInfo: ScriptInfo, changes: Iterable<TextChange>) {
+        
         for (const change of changes) {            
             scriptInfo.editContent(change.span.start, change.span.start + change.span.length, change.newText);
         }